requirements/base.txt
Django==3.2.9
django-environ==0.8.1
psycopg2==2.9.2
djangorestframework==3.12.4

celery==5.2.1
django-celery-results==2.2.0
django-celery-beat==2.2.1

whitenoise==5.3.0

django-filter==21.1
django-cors-headers==3.10.0

boto3==1.20.20
django-storages==1.12.3
